Add chatlist add, delete button
