+++ src/components/ChatList.tsx
@@ -1,26 +1,40 @@
-import React from 'react';
-import '../ChatList.css';
+// ChatList.tsx
 
-interface ChatListProps {
-  chats: string[];
-  handleChatClick: (chatName: string) => void;
-}
+import React, { useState } from 'react';
+import Chat from './Chat';
 
-const ChatList: React.FC<ChatListProps> = ({ chats, handleChatClick }) => {
+const ChatList: React.FC = () => {
+  const [chats, setChats] = useState<string[]>(['Chat 1', 'Chat 2', 'Chat 3', 'Chat 4']);
+  const [selectedChat, setSelectedChat] = useState<string | null>(null);
+
+  const handleChatClick = (chatName: string) => {
+    setSelectedChat(chatName);
+  };
+
+  const handleAddChat = () => {
+    const newChatName = `Chat ${chats.length + 1}`;
+    setChats((prevChats) => [...prevChats, newChatName]);
+  };
+
+  const handleRemoveChat = (chatName: string) => {
+    setChats((prevChats) => prevChats.filter((chat) => chat !== chatName));
+    setSelectedChat(null);
+  };
+
   return (
-    <div className="chat-list-container">
-      <h2>채팅 목록</h2>
-      <ul className="chat-list">
+    <div className="chat-list">
+      <div className="chat-sidebar">
+        <button className="chat-item" onClick={handleAddChat}>채팅 추가</button>
         {chats.map((chatName) => (
-          <li
-            key={chatName}
-            className="chat-list-item"
-            onClick={() => handleChatClick(chatName)}
-          >
+          <div key={chatName} className="chat-item" onClick={() => handleChatClick(chatName)}>
             {chatName}
-          </li>
+            <button onClick={() => handleRemoveChat(chatName)}>삭제</button>
+          </div>
         ))}
-      </ul>
+      </div>
+      <div className="chat-content">
+        {selectedChat && <Chat chatName={selectedChat} />}
+      </div>
     </div>
   );
 };
